LayoutTests/imported/w3c/web-platform-tests/focus/activeelement-after-focusing-different-site-iframe-contentwindow-expected.txt


FAIL Check trailing events assert_false: Should have gotten innerbodyfocus after willspineventloop expected false got true
PASS Check more trailing events
FAIL Check result assert_equals: Check log expected "outeronload,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:IFRAME,willbluriframe,didbluriframe,activeElement:IFRAME,willspineventloop,innerbodyfocus,innerbodyblur,innerbodyfocus," but got "outeronload,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:BODY,willbluriframe,didbluriframe,activeElement:BODY,willspineventloop,innerbodyfocus,innerbodyblur,innerbodyfocus,"
